fix: modify front-end adaptation
