+++ object/saml.go
@@ -189,113 +189,3 @@ func GetSamlResponse(application *Application, user *User, samlRequest string, h
 	res := base64.StdEncoding.EncodeToString([]byte(xmlStr))
 	return res, authnRequest.AssertionConsumerServiceURL, nil
 }
-
-type IdpEntityDescriptor struct {
-	XMLName  xml.Name `xml:"EntityDescriptor"`
-	DS       string   `xml:"xmlns:ds,attr"`
-	XMLNS    string   `xml:"xmlns,attr"`
-	MD       string   `xml:"xmlns:md,attr"`
-	EntityId string   `xml:"entityID,attr"`
-
-	IdpSSODescriptor IdpSSODescriptor `xml:"IDPSSODescriptor"`
-}
-
-type KeyInfo struct {
-	XMLName  xml.Name `xml:"http://www.w3.org/2000/09/xmldsig# KeyInfo"`
-	X509Data X509Data `xml:",innerxml"`
-}
-
-type X509Data struct {
-	XMLName         xml.Name        `xml:"http://www.w3.org/2000/09/xmldsig# X509Data"`
-	X509Certificate X509Certificate `xml:",innerxml"`
-}
-
-type X509Certificate struct {
-	XMLName xml.Name `xml:"http://www.w3.org/2000/09/xmldsig# X509Certificate"`
-	Cert    string   `xml:",innerxml"`
-}
-
-type KeyDescriptor struct {
-	XMLName xml.Name `xml:"KeyDescriptor"`
-	Use     string   `xml:"use,attr"`
-	KeyInfo KeyInfo  `xml:"KeyInfo"`
-}
-
-type IdpSSODescriptor struct {
-	XMLName                    xml.Name `xml:"urn:oasis:names:tc:SAML:2.0:metadata IDPSSODescriptor"`
-	ProtocolSupportEnumeration string   `xml:"protocolSupportEnumeration,attr"`
-	SigningKeyDescriptor       KeyDescriptor
-	NameIDFormats              []NameIDFormat      `xml:"NameIDFormat"`
-	SingleSignOnService        SingleSignOnService `xml:"SingleSignOnService"`
-	Attribute                  []Attribute         `xml:"Attribute"`
-}
-
-type NameIDFormat struct {
-	XMLName xml.Name
-	Value   string `xml:",innerxml"`
-}
-
-type SingleSignOnService struct {
-	XMLName  xml.Name
-	Binding  string `xml:"Binding,attr"`
-	Location string `xml:"Location,attr"`
-}
-
-type Attribute struct {
-	XMLName      xml.Name
-	Name         string `xml:"Name,attr"`
-	NameFormat   string `xml:"NameFormat,attr"`
-	FriendlyName string `xml:"FriendlyName,attr"`
-	Xmlns        string `xml:"xmlns,attr"`
-}
-
-func GetSamlMeta(application *Application, host string) (*IdpEntityDescriptor, error) {
-	//_, originBackend := getOriginFromHost(host)
-	cert := getCertByApplication(application)
-	block, _ := pem.Decode([]byte(cert.PublicKey))
-	publicKey := base64.StdEncoding.EncodeToString(block.Bytes)
-
-	origin := beego.AppConfig.String("origin")
-	_, originBackend := getOriginFromHost(host)
-	if origin != "" {
-		originBackend = origin
-	}
-	d := IdpEntityDescriptor{
-		XMLName: xml.Name{
-			Local: "md:EntityDescriptor",
-		},
-		DS:       "http://www.w3.org/2000/09/xmldsig#",
-		XMLNS:    "urn:oasis:names:tc:SAML:2.0:metadata",
-		MD:       "urn:oasis:names:tc:SAML:2.0:metadata",
-		EntityId: originBackend,
-		IdpSSODescriptor: IdpSSODescriptor{
-			SigningKeyDescriptor: KeyDescriptor{
-				Use: "signing",
-				KeyInfo: KeyInfo{
-					X509Data: X509Data{
-						X509Certificate: X509Certificate{
-							Cert: publicKey,
-						},
-					},
-				},
-			},
-			NameIDFormats: []NameIDFormat{
-				{Value: "urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress"},
-				{Value: "urn:oasis:names:tc:SAML:2.0:nameid-format:persistent"},
-				{Value: "urn:oasis:names:tc:SAML:2.0:nameid-format:transient"},
-			},
-			Attribute: []Attribute{
-				{Xmlns: "urn:oasis:names:tc:SAML:2.0:assertion", Name: "Email", NameFormat: "urn:oasis:names:tc:SAML:2.0:attrname-format:basic", FriendlyName: "E-Mail"},
-				{Xmlns: "urn:oasis:names:tc:SAML:2.0:assertion", Name: "DisplayName", NameFormat: "urn:oasis:names:tc:SAML:2.0:attrname-format:basic", FriendlyName: "displayName"},
-				{Xmlns: "urn:oasis:names:tc:SAML:2.0:assertion", Name: "Name", NameFormat: "urn:oasis:names:tc:SAML:2.0:attrname-format:basic", FriendlyName: "Name"},
-			},
-			SingleSignOnService: SingleSignOnService{
-				Binding:  "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect",
-				Location: fmt.Sprintf("%s/login/oauth/authorize", host),
-			},
-			ProtocolSupportEnumeration: "urn:oasis:names:tc:SAML:2.0:protocol",
-		},
-	}
-
-	return &d, nil
-}
